Remove selection set API from read
